% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aw.R
\docType{data}
\name{aw}
\alias{aw}
\title{Data of aw}
\format{
A data frame with 9 rows and 2 variables:
\describe{
\item{aw}{Water activity}
\item{sqrtGR}{Square root of the growth rate}
\item{GR}{Growth rate}
}
}
\usage{
data(aw)
}
\description{
A dataset containing ???
}
\keyword{datasets}
