% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baranyi_full.R
\name{BaranyiFM}
\alias{BaranyiFM}
\title{Baranyi and Roberts full growth model}
\usage{
BaranyiFM(t, Y0, Ymax, MUmax, lag)
}
\arguments{
\item{t}{is a numeric vector indicating the time of the experiment}

\item{Y0}{is the natural logarithm of the initial microbial concentration (\code{ln(N0)}) at time=0}

\item{Ymax}{is the natural logarithm of the maximum concentration (\code{ln(Nmax)}) reached by the microorganism}

\item{MUmax}{is the maximum specific growth rate given in time units}

\item{lag}{is the duration of the lag phase in time units}
}
\value{
An object of nls class with the fitted parameters of the model
}
\description{
\code{BaranyiFM} function to fit the Baranyi & Roberts full growth model to a complete microbial growth curve.
Returns the model parameters estimated according to data collected in microbial growth experiments.
}
\details{
The model's inputs are:

\code{t}: time, assuming time zero as the beginning of the experiment.

\code{Y(t)}: the natural logarithm of the microbial concentration (\code{ln(N(t)}) measured at time t.

Users should make sure that the microbial concentration input is entered in natural logarithm, \code{Y(t) = ln(N(t))}.
}
\examples{
library(gslnls)
data(growthfull)
initial_values = list(Y0=-0.1, Ymax=22, MUmax=1.7, lag=5)
fit <- gsl_nls(lnN ~ BaranyiFM(Time, Y0, Ymax, MUmax, lag),
               data=growthfull,
               start =  initial_values)
summary(fit)

}
\references{
\insertRef{Baranyi1994}{predmicror}
}
\author{
Vasco Cadavez \email{vcadavez@ipb.pt} and Ursula Gonzales-Barron \email{ubarron@ipb.pt}
}
\keyword{Baranyi}
\keyword{Roberts}
\keyword{full}
\keyword{model}
