% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baranyi_red.R
\name{BaranyiRM}
\alias{BaranyiRM}
\title{Baranyi and Roberts reduced growth model}
\usage{
BaranyiRM(t, Y0, MUmax, lag)
}
\arguments{
\item{t}{is a numeric vector indicating the time of the experiment}

\item{Y0}{is the natural logarithm of the initial microbial concentration (\code{ln(N0)}) at time=0}

\item{MUmax}{is the maximum specific growth rate given in time units}

\item{lag}{is the duration of the lag phase in time units}
}
\value{
An object of nls class with the fitted parameters of the model
}
\description{
\code{BaranyiRM} function to fit the Baranyi and Roberts growth model to a reduced microbial growth curve.
Returns the model parameters estimated according to data collected in microbial growth experiments.
}
\details{
Model's inputs are:

\code{t}: time, assuming time zero as the beginning of the experiment.

\code{Y(t)}: the natural logarithm of the microbial concentration (\code{ln(N(t)}) measured at time t.

Users should make sure that the microbial concentration input is entered in natural logarithm, \code{Y(t) = ln(N(t))}.
}
\examples{
## Example: Baranyi reduced model
library(gslnls)
data(growthred)  # simulated data set.
initial_values = list(Y0=0.1, MUmax=1.7, lag=5) # define the initial values
fit <- gsl_nls(lnN ~ BaranyiRM(Time, Y0, MUmax, lag),
           data=growthred,
           start =  initial_values)
summary(fit)

}
\references{
\insertRef{Baranyi1995}{predmicror}
}
\author{
Vasco Cadavez, \email{vcadavez@ipb.pt} and Ursula Gonzales-Barron, \email{ubarron@ipb.pt}
}
\keyword{Baranyi}
\keyword{Reduced}
\keyword{growth}
