% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omnibus.R
\name{bias_factor}
\alias{bias_factor}
\alias{accuracy_factor}
\title{Bias and accuracy factors}
\usage{
bias_factor(observed, predicted)

accuracy_factor(observed, predicted)
}
\arguments{
\item{observed, predicted}{Numeric vectors. Values must be positive.}
}
\value{
A numeric scalar.
}
\description{
\code{bias_factor()} and \code{accuracy_factor()} calculate the multiplicative model
bias and accuracy factors commonly used in predictive microbiology model
validation.
}
\examples{
observed <- c(7.0, 6.0, 5.0)
predicted <- c(7.1, 5.9, 5.2)
bias_factor(observed, predicted)
accuracy_factor(observed, predicted)
}
