% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bixina.R
\docType{data}
\name{bixina}
\alias{bixina}
\title{Data concerning \emph{Staphylococcus aureus} microbial inactivation in beef}
\format{
A data frame with 18 rows and 3 variables:
\describe{
\item{Time}{Time in minutes}
\item{Rep}{Repetition}
\item{lnN}{Number of microrganism in ln scale}
}
}
\usage{
data(bixina)
}
\description{
A dataset containing time (hours), repetition and number of microrganisms (ln N).
}
\keyword{beef}
\keyword{bixina}
\keyword{inactivation}
