% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardinal_aw.R
\name{CMAW}
\alias{CMAW}
\title{Cardinal model for water activity}
\usage{
CMAW(x, AWmin, MUopt, AWopt)
}
\arguments{
\item{x}{is a numeric vector indicating the water activity of the experiment}

\item{AWmin}{is minimum water activity for growth}

\item{MUopt}{is the optimum growth rate}

\item{AWopt}{is optimum water activity for growth}
}
\value{
An object of nls class with the fitted parameters of the model
}
\description{
\code{CMAW} function to fit the water activity cardinal model (Rosso et al., 1993).
Returns the model parameters estimated according to data collected in microbial growth experiments.
}
\details{
The model's inputs are:

\code{x}: Water activity

\code{sqrtGR}: the square root of the growth rate ($h^{-1}$)

Users should make sure that the growth rate input is entered after a square root transformation, \code{sqrGR = sqrt(GR)}.
}
\examples{
library(gslnls)
data(aw)
initial_values = list(AWmin=0.89, MUopt=1.0, AWopt=0.98)
fit <- gsl_nls(sqrtGR ~ CMAW(aw,AWmin,MUopt,AWopt),
               data=aw,
               start =  initial_values)
summary(fit)

}
\references{
\insertRef{Rosso1995}{predmicror}
}
\author{
Vasco Cadavez \email{vcadavez@ipb.pt} and Ursula Gonzales-Barron \email{ubarron@ipb.pt}
}
\keyword{Cardinal}
\keyword{Rosso}
\keyword{activity}
\keyword{water}
