% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardinal_inh.R
\name{CMInh}
\alias{CMInh}
\title{Cardinal model for growth inhibitors}
\usage{
CMInh(x, MIC, MUopt, alpha)
}
\arguments{
\item{x}{is a numeric vector indicating the inhibitor concentration of the experiment}

\item{MIC}{is the minimum inhibitory concentration (mM or \%, accordingly)}

\item{MUopt}{is the optimum growth rate}

\item{alpha}{is the shape parameter of the curve (alpha = 1 the shape is linear;
alpha > 1 the shape is downward concave; and alpha < 1 the shape is upward concave)}
}
\value{
An object of nls class with the fitted parameters of the model
}
\description{
\code{CMInh} function to fit the growth inhibitors cardinal model (Rosso et al, 1993).
Returns the model parameters estimated according to data collected in microbial growth experiments.
}
\details{
The model's inputs are:

\code{x}: growth inhibitor concentration

\code{sqrtGR}: the square root of the growth rate ($time^{-1}$)

Users should make sure that the growth rate input is entered after a square root transformation, $sqrGR = sqrt(GR)$.
}
\examples{
library(gslnls)
data(inh)
initial_values <- list(MIC = 0.89, MUopt = 1.0, alpha = 1)
fit <- gsl_nls(sqrtGR ~ CMInh(Conce, MIC, MUopt, alpha),
  data = inh,
  start = initial_values
)
summary(fit)

}
\references{
\insertRef{Rosso1995}{predmicror}
}
\author{
Vasco Cadavez \email{vcadavez@ipb.pt} and Ursula Gonzales-Barron \email{ubarron@ipb.pt}
}
\keyword{Cardinal}
\keyword{Inhibitor}
\keyword{Rosso}
