% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardinal_ph.R
\name{CMPH}
\alias{CMPH}
\title{Cardinal model for pH}
\usage{
CMPH(x, pHmax, pHmin, MUopt, pHopt)
}
\arguments{
\item{x}{is a numeric vector indicating the pH of the experiment}

\item{pHmax}{is the maximum pH for growth}

\item{pHmin}{is the minimum pH for growth}

\item{MUopt}{is the optimum growth rate}

\item{pHopt}{is the optimum pH for growth}
}
\value{
An object of nls class with the fitted parameters of the model
}
\description{
\code{CMPH} function to fit the pH cardinal model (Rosso et al, 1993).
Returns the model parameters estimated according to data collected in microbial growth experiments.
}
\details{
The model's inputs are:

\code{x}: pH

\code{sqrtGR}: the square root of the growth rate ($time^{-1}$)

Users should make sure that the growth rate input is entered after a square root transformation, \code{sqrGR = sqrt(GR)}.
}
\examples{
library(gslnls)
data(ph)
initial_values = list(pHmax=9, pHmin=3, MUopt=1.0, pHopt=7)
fit <- gsl_nls(sqrtGR ~ CMPH(pH,pHmax,pHmin,MUopt,pHopt),
               data=ph,
               start =  initial_values)
summary(fit)

}
\references{
\insertRef{Rosso1995}{predmicror}
}
\author{
Vasco Cadavez \email{vcadavez@ipb.pt} and Ursula Gonzales-Barron \email{ubarron@ipb.pt}
}
\keyword{Cardinal}
\keyword{Rosso}
\keyword{Temperature}
