% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardinal_temp.R
\name{CMTI}
\alias{CMTI}
\title{Cardinal model for temperature}
\usage{
CMTI(x, Tmax, Tmin, MUopt, Topt)
}
\arguments{
\item{x}{is a numeric vector indicating the temperature of the experiment}

\item{Tmax}{maximum temperature for growth}

\item{Tmin}{is minimum temperature for growth}

\item{MUopt}{is the optimum growth rate}

\item{Topt}{is optimum temperature for growth}
}
\value{
An object of nls class with the fitted parameters of the model
}
\description{
\code{CMTI} function to fit the temperature cardinal model (Rosso et al, 1993).
Returns the model parameters estimated according to data collected in microbial growth experiments.
}
\details{
The model's inputs are:

\code{x}: Temperature

\code{sqrtGR}: the square root of the growth rate ($h^{-1}$)

Users should make sure that the growth rate input is entered after a square root transformation, \code{sqrGR = sqrt(GR)}.
}
\examples{
library(gslnls)
data(salmonella)
initial_values = list(Tmax=42, Tmin=1, MUopt=1.0, Topt=37)
fit <- gsl_nls(sqrtGR ~ CMTI(Temp,Tmax,Tmin,MUopt,Topt),
               data=salmonella,
               start =  initial_values)
summary(fit)

plot(salmonella$Temp, salmonella$sqrtGR^2)
lines(salmonella$Temp, fitted(fit)^2, col="green")
plot(salmonella$Temp, salmonella$sqrtGR)
lines(salmonella$Temp, fitted(fit), col="red")

}
\references{
\insertRef{Rosso1995}{predmicror}
}
\author{
Vasco Cadavez \email{vcadavez@ipb.pt} and Ursula Gonzales-Barron \email{ubarron@ipb.pt}
}
\keyword{Cardinal}
\keyword{Rosso}
\keyword{Temperature}
