% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfull.R
\docType{data}
\name{growthfull}
\alias{growthfull}
\title{Data of a complete curve of microbial growth}
\format{
A data frame with 13 rows and 3 variables.
\describe{
\item{Time}{Time in minutes}
\item{logN}{Number of microrganism in log10 scale}
\item{lnN}{Number of microrganism in ln scale}
}
}
\usage{
data(growthfull)
}
\description{
A dataset containing simulated data for a full growth model
}
\keyword{datasets}
