% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthred.R
\docType{data}
\name{growthred}
\alias{growthred}
\title{Data of a reduced curve of microbial growth}
\format{
A data frame with 9 rows and 3 variables.
\describe{
\item{Time}{Time in minutes}
\item{logN}{Number of microrganism in log10 scale}
\item{lnN}{Number of microrganism in ln scale}
}
}
\usage{
data(growthred)
}
\description{
A dataset containing simulated data for a reduced growth model
}
\keyword{datasets}
