% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inh.R
\docType{data}
\name{inh}
\alias{inh}
\title{Data of INH antimicrobials}
\format{
A data frame with 8 rows and 3 variables.
\describe{
\item{Conce}{Antimicrobial concentration}
\item{sqrtGR}{Square root of the growth rate}
\item{GR}{Growth rate}
}
}
\usage{
data(inh)
}
\description{
A dataset containing antimicrobial concentration and growth rate data
}
\keyword{datasets}
