% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mafart2005Li11.R
\docType{data}
\name{mafart2005Li11}
\alias{mafart2005Li11}
\title{Data of microbial inactivation Albert and Mafart (2005)}
\format{
A data frame with 10 rows and 2 variables.
\describe{
\item{Time}{Time}
\item{logN}{Number of microrganism in log10 scale}
}
}
\usage{
data(mafart2005Li11)
}
\description{
A dataset containing time and log10 counts for microbial inactivation
}
\keyword{datasets}
