% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph.R
\docType{data}
\name{ph}
\alias{ph}
\title{Data pH}
\format{
A data frame with 14 rows and 2 variables.
\describe{
\item{pH}{pH of food}
\item{sqrtGR}{Square root of the growth rate}
\item{GR}{Growth rate}
}
}
\usage{
data(ph)
}
\description{
A dataset containing ???
}
\keyword{datasets}
