% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant_app.R
\name{predmicror_assistant_app}
\alias{predmicror_assistant_app}
\title{Launch the predmicror assistant Shiny app}
\usage{
predmicror_assistant_app()
}
\value{
Runs the Shiny app.
}
\description{
Starts the local assistant app bundled with the package.
}
\examples{
\dontrun{
predmicror_assistant_app()
}

}
