% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richards_nolag.R
\name{RichardsNLM}
\alias{RichardsNLM}
\title{Richards no lag growth model}
\usage{
RichardsNLM(t, Y0, Ymax, MUmax, m = 1)
}
\arguments{
\item{t}{is a numeric vector indicating the time of the experiment}

\item{Y0}{is the natural logarithm of the initial microbial concentration (\code{ln(N0)}) at time=0}

\item{Ymax}{is the natural logarithm of the maximum concentration (\code{ln(Nmax)}) reached by the microorganism}

\item{MUmax}{is the maximum specific growth rate given in time units}

\item{m}{is the shape parameter of the Richards model (default = 1)}
}
\value{
A numeric vector with the fitted values
}
\description{
\code{RichardsNLM} function to fit the Richards no lag growth model to an incomplete microbial growth curve.
Returns the model parameters estimated according to data collected in microbial growth experiments.
}
\details{
Model's inputs are:

\code{t}: time, assuming time zero as the beginning of the experiment.

\code{Y(t)}: the natural logarithm of the microbial concentration (\code{ln(N(t)}) measured at time t.

Users should make sure that the microbial concentration input is entered in natural logarithm, \code{Y(t) = ln(N(t))}.
}
\examples{
## Example: Richards no lag model
library(gslnls)
data(growthnolag) # simulated data set.
initial_values <- list(Y0 = 0, Ymax = 22, MUmax = 1.7)
## Fitting the function to the experimental data
fit <- gsl_nls(lnN ~ RichardsNLM(Time, Y0, Ymax, MUmax),
  data = growthnolag,
  start = initial_values
)
summary(fit)

}
\references{
\insertRef{Richards1959}{predmicror}
}
\author{
Vasco Cadavez, \email{vcadavez@ipb.pt} and Ursula Gonzales-Barron, \email{ubarron@ipb.pt}
}
\keyword{Growth}
\keyword{Richards}
\keyword{nolagmodel}
