% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.R
\docType{data}
\name{salmonella}
\alias{salmonella}
\title{Potential growth of Salmonella typhimurium on cooked chicken}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{Temp}{Temperature (ºC)}
\item{GR}{Specific growth rate (log10/h)}
\item{sqrtGR}{Square root of the specific growth rate (GR)}
}
}
\usage{
data(salmonella)
}
\description{
A dataset containing the specific growth rate of Salmonela typhimurium on cooked chicken
}
\references{
\insertRef{Oscar2002}{predmicror}
}
\keyword{datasets}
