% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_inactivation_mafart.R
\name{WeibullM}
\alias{WeibullM}
\title{Weibull inactivation model Mafart}
\usage{
WeibullM(x, Y0, sigma, alpha)
}
\arguments{
\item{x}{is a numeric vector indicating the heating time under a constant temperature of the experiment}

\item{Y0}{is the natural logarithm of the initial (time=0) bacterial concentration (N0)}

\item{sigma}{is the time of first decimal reduction}

\item{alpha}{which is a shape parameter}
}
\value{
A numeric vector with the fitted values
}
\description{
\code{WeibullM} inactivation model for microbial inactivation curve.
Returns the model parameters estimated according to data collected in microbial inactivation experiments.
}
\details{
The model's inputs are:

\code{t}: time, assuming time zero as the beginning of the experiment.

\code{Y(t)}: the natural logarithm of the bacterial concentration (\code{ln(N(t))}) measured at time t.

Users should make sure that the bacterial concentration input is entered
in natural logarithm, \code{Y(t) = ln(N(t))}.
}
\examples{
library(gslnls)
data(bixina)
initial_values <- list(Y0 = 5.75, sigma = 12.8, alpha = 2.4)
fit <- gsl_nls(lnN ~ WeibullM(Time, Y0, sigma, alpha),
  data = bixina,
  start = initial_values
)
summary(fit)

plot(lnN ~ Time, data = bixina)
lines(bixina$Time, predict(fit), col = "blue")

}
\references{
\Sexpr[results=rd,stage=build]{Rdpack::insert_ref(key="Mafart2002",package="predmicror")}
}
\author{
Vasco Cadavez \email{vcadavez@ipb.pt} and Ursula Gonzales-Barron \email{ubarron@ipb.pt}
}
\keyword{Inactivation}
\keyword{Mafart}
\keyword{Weibull}
