% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_inactivation_modified_mafart.R
\name{WeibullMM}
\alias{WeibullMM}
\title{Weibull inactivation modified model Mafart}
\usage{
WeibullMM(x, Y0, Yres, sigma, alpha)
}
\arguments{
\item{x}{is a numeric vector indicating the heating time under a constant temperature of the experiment}

\item{Y0}{is the log10 of the initial bacterial concentration (at time t=0)}

\item{Yres}{is the log10 of the residual bacterial concentration (at the end of the experiment)}

\item{sigma}{represents the time of the first decimal reduction concentration
for the part of the population not belonging to Yres}

\item{alpha}{is the shape parameter and allows to catch the curve concavity or convexity}
}
\value{
An object of nls class with the fitted parameters of the model
}
\description{
\code{WeibullMM} inactivation model for microbial inactivation curve.
Returns the model parameters estimated according to data collected in microbial inactivation experiments.
}
\details{
The model's inputs are:

\code{t}: time, assuming time zero as the beginning of the experiment.

\code{Y(t)}: the bacterial concentration ($Y(t)$) measured at time t.

Users should make sure to use the base 10 logarithm bacterial concentration (Y(t)) as input.
}
\examples{
library(gslnls)
data(bixina)

}
\references{
\insertRef{Mafart2005}{predmicror}
}
\author{
Vasco Cadavez \email{vcadavez@ipb.pt} and Ursula Gonzales-Barron \email{ubarron@ipb.pt}
}
\keyword{Inactivation}
\keyword{Mafart}
\keyword{Modified}
\keyword{Weibull}
