% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_inactivation_peleg_huang.R
\name{WeibullPH}
\alias{WeibullPH}
\title{Weibull inactivation model Peleg and Huang}
\usage{
WeibullPH(t, Y0, k, alpha)
}
\arguments{
\item{t}{is a numeric vector indicating the time of the experiment}

\item{Y0}{is the natural logarithm of the initial (time=0) bacterial concentration}

\item{k}{is the inactivation rate (ln units/time)}

\item{alpha}{is the shape parameter}
}
\value{
A numeric vector with the fitted values
}
\description{
\code{WeibullPH} inactivation model for microbial inactivation curve.
Returns the model parameters estimated according to data collected in microbial inactivation experiments.
}
\details{
The model's inputs are:

t: time, assuming time zero as the beginning of the experiment.

Y(t): the natural logarithm of the bacterial concentration X(t) measured at time t.

Users should make sure that the bacterial concentration input is entered in natural logarithm, Y(t) = ln(X(t)).

The following parameters can be estimated using Weibull function:

\code{t}: is heating time under a constant temperature

\code{Y0}: is the initial (time=0) bacterial counts in natural logarithm of the initial bacterial counts;

\code{k}: is the inactivation rate (ln units/time)

\code{alpha}: is the shape parameter of the survival curve
}
\examples{
library(gslnls)
data(bixina)
initial_values <- list(Y0 = 6.0, k = 1.0, alpha = 0.2)
fit <- gsl_nls(lnN ~ WeibullPH(Time, Y0, k, alpha),
  data = bixina,
  start = initial_values
)
summary(fit)

plot(lnN ~ Time, data = bixina)
lines(bixina$Time, predict(fit), col = "blue")

}
\references{
\Sexpr[results=rd,stage=build]{Rdpack::insert_ref(key="Huang2009Inac",package="predmicror")}
}
\author{
Vasco Cadavez \email{vcadavez@ipb.pt} and Ursula Gonzales-Barron \email{ubarron@ipb.pt}
}
\keyword{Inactivation}
\keyword{Weibull}
