% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bixina.R
\docType{data}
\name{bixina}
\alias{bixina}
\title{Data of microbial inactivation}
\format{
A data frame with 10 rows and 3 variables.
\describe{
\item{Time}{Time in minutes}
\item{Rep}{Repetition}
\item{lnN}{Number of microrganism in ln scale}
}
}
\usage{
data(bixina)
}
\description{
A dataset containing
}
\keyword{datasets}
