% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geeraerd_inactivation.R
\name{GeeraerdST}
\alias{GeeraerdST}
\title{Geeraerd inactivation model}
\usage{
GeeraerdST(x, Y0, Yres, kmax, Sl)
}
\arguments{
\item{x}{is a numeric vector indicating the heating time under a constant temperature of the experiment}

\item{Y0}{is the initial (time=0) bacterial concentration (log10(N0))}

\item{Yres}{is a low asymptote reflecting the presence of a resistant sub-population (log10(Nres))}

\item{kmax}{is the maximum inactivation rate}

\item{Sl}{represents shoulder phase preceding the sharp inactivation slope of the curve}
}
\value{
An object of nls class with the fitted parameters of the model
}
\description{
\code{GeeraerdST} inactivation model for microbial inactivation curve.
Returns the model parameters estimated according to data collected in microbial inactivation experiments.
}
\details{
The model's inputs are:

\code{t}: time, assuming time zero as the beginning of the experiment.

\code{N(t)}: the bacterial concentration measured at time t.

Users should make sure that the bacterial concentration input is entered
in base 10 logarithm, \code{Y(t) = log10(N(t))}.
}
\examples{
library(gslnls)
data(mafart2005Li11)
initial_values <- list(Y0 = 10, Yres = 7, kmax = 0.7, Sl = 4)
fit <- gsl_nls(logN ~ GeeraerdST(Time, Y0, Yres, kmax, Sl),
  data = mafart2005Li11,
  start = initial_values
)
summary(fit)

plot(logN ~ Time, data = mafart2005Li11)
lines(mafart2005Li11$Time, predict(fit), col = "blue")

}
\references{
\insertRef{Geeraerd2005}{predmicror}
}
\author{
Vasco Cadavez \email{vcadavez@ipb.pt} and Ursula Gonzales-Barron \email{ubarron@ipb.pt}
}
\keyword{Geeraerd}
\keyword{Inactivation}
\keyword{Shoulder}
\keyword{Tail}
