% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predmicror-package.R
\docType{package}
\name{predmicror-package}
\alias{predmicror}
\alias{predmicror-package}
\title{predmicror: predmicror: Fitting predictive microbiology models in r}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A package for fitting predictive microbiology models to experimental data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fsqanalytics/predmicror}
  \item Report bugs at \url{https://github.com/fsqanalytics/predmicror/issues}
}

}
\author{
\strong{Maintainer}: Vasco Cadavez \email{vcadavez@ipb.pt} (\href{https://orcid.org/0000-0002-3077-7414}{ORCID})

Authors:
\itemize{
  \item Ursula Gonzales-Barron \email{ubarron@ipb.pt} (\href{https://orcid.org/0000-0002-8462-9775}{ORCID})
}

}
\keyword{internal}
