% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_inactivation_peleg_huang.R
\name{WeibullPH}
\alias{WeibullPH}
\title{Weibull inactivation model Peleg and Huang}
\usage{
WeibullPH(t, Y0, k, alpha)
}
\arguments{
\item{t}{is a numeric vector indicating the time of the experiment}

\item{Y0}{is the base 10 logarithm of the initial (time=0) bacterial concentration}

\item{k}{is the inactivation rate ()}

\item{alpha}{is the ???}
}
\value{
An object of nls class with the fitted parameters of the model
}
\description{
\code{WeibullPH} inactivation model for microbial inactivation curve.
Returns the model parameters estimated according to data collected in microbial inactivation experiments.
}
\details{
The model's inputs are:

t: time, assuming time zero as the beginning of the experiment.

Y(t): the base 10 logarithm of the bacterial concentration X(t) measured at time t.

Users should make sure that the bacterial concentration input is entered in base 10 logarithm, Y(t) = log10(X(t)).

The following parameters can be estimated using Weibull function:

\code{t}: is heating time under a constant temperature

\code{Y0}: is the initial (time=0) bacterial counts in base 10 logarithm of the initial bacterial counts;

\code{k}: is the inactivation rate (log cfu/s or log cfu/min, or log cfu/h)

\code{alpha}: is the ????
}
\examples{
library(gslnls)
data(bixina)
initial_values = list(Y0=6.0, k=1.0, alpha=0.2)
fit <- gsl_nls(lnN ~ WeibullPH(Time, Y0, k, alpha),
           data=bixina,
           start =  initial_values)
summary(fit)

plot(lnN ~ Time, data=bixina)
lines(bixina$Time, predict(fit), col="blue")

}
\references{
\insertRef{Huang2009Inac}{predmicror}
}
\author{
Vasco Cadavez \email{vcadavez@ipb.pt} and Ursula Gonzales-Barron \email{ubarron@ipb.pt}
}
\keyword{Inactivation}
\keyword{Weibull}
